Implement Average Damage Pipeline
Implemented the processing of the average damage pipeline. It analyses
what the average damage will be over an infinite number of samples.
I still need to add status effects and the damage that they deal, but
this at least provides an estimate for the average damage per trigger
pull.

In the Weapon class there is also functions for getting the average
burst DPS and sustained DPS for comparisons between weapons.