Include Status Effect damage in calculations
Added processing of status effect damage to the DamagePipeline. Fire and
GetAverageDamagePerShot now return a pair of floats depicting the direct
damage per shot and DOT damage per shot to allow analysis of status
effect damage and relative contributions compared to the direct damage.

The current state of the system does not account for damage boosting
status effects (e.g. Viral, Magnetic, Corrosive, Cold, Puncture, Tau),
as it would require estimations of total uptime from the whole attack
and all sub-attacks. It may be possible to implement such a feature,
but it would have to be locked behind a user-passed setting to ensure
the final results are properly understood.