Create ModFactory GetMod method
Created a GetMod method which retrieves mod data from the database
to create a new mod instance with all associated details filled. This is
the first of the data-based approaches to object creation. Some demo
database entries are provided in the data/Lotus folder.

There are some limitations of the ModFactory method as it stands.
Currently it does not process conditional requirements so all mod
effects are applied permanently. Some changes will be required to the
Mod structure to ensure the data format is as close to the original
game's structure before this conditional parsing may be implemented.